move column implementation to column.go
